Improve OTP handling in TOTP and Entry modules
This commit enhances the OTP handling in both TOTP and Entry modules. TOTPAlgorithm was updated to handle UpperCase characters. Display traits were implemented for TOTPAlgorithm, TOTP, and OTPCode to allow pretty-printing. Public accessibility was added to the fields of the TOTP struct. A new convenience method set_otp was introduced in the Entry module for setting TOTP entries.